show version in page title

+++ python/video.py
@@ -44,6 +44,8 @@ def getVideoInfo(videoPath, pipeIn, width, height, frameRate):
 
     while matchInfo or matchFrame:
       line = pipeIn.stderr.readline()
+      if type(line) != str:
+        line = str(line, 'utf-8', errors='ignore')
       sys.stdout.write(line)
       if not line:
         break
@@ -110,7 +112,7 @@ def prepare(video, steps):
   decodec = optDecode['codec'] if 'codec' in optDecode else config.defaultDecodec  # pylint: disable=E1101
   optRange = steps[1]
   start = int(optRange['start']) if 'start' in optRange else 0
-  outDir = config.outDir  # pylint: disable=E1101
+  outDir, uploadDir = config.outDir, config.uploadDir  # pylint: disable=E1101
   procSteps = stepVideo + list(steps[2:-1])
   process, nodes = genProcess(procSteps)
   root = begin(Node({'op': 'video', 'encodec': encodec}, 1, 2, 0), nodes, False)
@@ -131,17 +133,21 @@ def prepare(video, steps):
     stop = -1
   root.total = -1 if stop < 0 else stop - start
   outputPath = optEncode.get('file', '') or outDir + '/' + config.getPath()
+  t0Path = uploadDir + '/t0.ts'
+  t1Path = uploadDir + '/t1.ts'
   commandIn = [
     ffmpegPath,
     '-i', video,
-    '-an',
-    '-sn',
+    '-vn',
+    '-c', 'copy',
+    t0Path,
+    '-map', '0:v',
     '-f', 'rawvideo',
     '-pix_fmt', pix_fmt]
   if len(decodec):
     commandIn.extend(decodec.split(' '))
   commandIn.append('-')
-  commandOut = [
+  commandVideo = [
     ffmpegPath,
     '-y',
     '-f', 'rawvideo',
@@ -149,22 +155,29 @@ def prepare(video, steps):
     '-s', '',
     '-r', '',
     '-i', '-',
-    '-i', video,
-    '-map', '0:v',
-    '-map', '1?',
-    '-map', '-1:v',
+    '-c:v:0'] + encodec.split(' ') + [t1Path]
+  commandOut = [
+    ffmpegPath,
+    '-y',
+    '-i', t1Path,
+    '-i', t0Path,
     '-c:1', 'copy',
-    '-c:v:0'] + encodec.split(' ')
+    '-c:0', 'copy',
+    outputPath
+  ]
   if start > 0:
-    commandOut = commandOut[:12] + commandOut[22:]
-  commandOut.append(outputPath)
+    commandIn = commandIn[:3] + commandIn[7:]
+    commandOut = commandOut[:4] + commandOut[8:]
+    files = (t1Path, t0Path)
+  else:
+    files = (t1Path,)
   frameRate = optEncode.get('frameRate', 0)
   width = optDecode.get('width', 0)
   height = optDecode.get('height', 0)
   sizes = filter((lambda opt: opt['op'] == 'SR' or opt['op'] == 'resize'), procSteps)
-  return outputPath, process, start, stop, root, commandIn, commandOut, slomos, sizes, width, height, frameRate
+  return outputPath, process, start, stop, root, commandIn, commandVideo, commandOut, files, slomos, sizes, width, height, frameRate
 
-def setupInfo(root, commandOut, slomos, sizes, start, width, height, frameRate, totalFrames):
+def setupInfo(root, commandVideo, slomos, sizes, start, width, height, frameRate, totalFrames):
   if root.total < 0 and totalFrames > 0:
     root.total = totalFrames - start
   if frameRate:
@@ -178,35 +191,36 @@ def setupInfo(root, commandOut, slomos, sizes, start, width, height, frameRate, 
     else: # resize
       outWidth = round(outWidth * opt['scaleW']) if 'scaleW' in opt else opt['width']
       outHeight = round(outHeight * opt['scaleH']) if 'scaleH' in opt else opt['height']
-  commandOut[7] = f'{outWidth}x{outHeight}'
-  commandOut[9] = str(frameRate)
+  commandVideo[7] = f'{outWidth}x{outHeight}'
+  commandVideo[9] = str(frameRate)
   root.multipleLoad(width * height * 3)
   initialETA(root)
   root.reset().trace(0)
 
 def SR_vid(video, by, *steps):
-  outputPath, process, start, stop, root, commandIn, commandOut, slomos, sizes, *info = prepare(video, steps)
+  outputPath, process, start, stop, root, commandIn, commandVideo, commandOut, files, slomos, sizes, *info = prepare(video, steps)
   if by:
     pipeIn = popen(commandIn)
     width, height, *more = getVideoInfo(video, pipeIn, *info)
   else:
     width, height, *more = getVideoInfo(video, False, *info)
     pipeIn = popen(commandIn)
-  setupInfo(root, commandOut, slomos, sizes, start, width, height, *more)
-  pipeOut = sp.Popen(commandOut, stdin=sp.PIPE, stdout=sp.PIPE, stderr=sp.PIPE, bufsize=bufsize, shell=True)
+  setupInfo(root, commandVideo, slomos, sizes, start, width, height, *more)
+  pipeVideo = sp.Popen(commandVideo, stdin=sp.PIPE, stdout=sp.PIPE, stderr=sp.PIPE, bufsize=bufsize, shell=True)
   def p(raw_image=None):
     bufs = process((raw_image, height, width))
     if (not bufs is None) and len(bufs):
       for buffer in bufs:
         if buffer:
-          pipeOut.stdin.write(buffer)
+          pipeVideo.stdin.write(buffer)
     if raw_image:
       root.trace()
 
   try:
-    createEnqueueThread(pipeOut.stdout, 0)
+    pipeOut = 0
+    createEnqueueThread(pipeVideo.stdout, 0)
     createEnqueueThread(pipeIn.stderr, 1)
-    createEnqueueThread(pipeOut.stderr, 1)
+    createEnqueueThread(pipeVideo.stderr, 1)
     i = 0
     while (stop < 0 or i <= stop) and not context.stopFlag.is_set():
       raw_image = pipeIn.stdout.read(width * height * pixBytes) # read width*height*6 bytes (= 1 frame)
@@ -219,13 +233,23 @@ def SR_vid(video, by, *steps):
       idle()
     p()
 
+    pipeVideo.communicate(timeout=300)
+    pipeOut = sp.Popen(commandOut, stdin=sp.PIPE, stdout=sp.PIPE, stderr=sp.PIPE, encoding='utf_8', errors='ignore')
+    createEnqueueThread(pipeOut.stdout, 0)
+    createEnqueueThread(pipeOut.stderr, 0)
     pipeOut.communicate(timeout=300)
   finally:
     pipeIn.terminate()
-    pipeOut.terminate()
+    pipeVideo.terminate()
+    if pipeOut:
+      pipeOut.terminate()
     clean()
     try:
-      os.remove(video)
+      if not by:
+        os.remove(video)
+      if start > 0:
+        os.remove(files[1])
+      os.remove(files[0])
     except:
       log.warning('Timed out waiting ffmpeg to terminate, need to remove {} manually.'.format(video))
   readSubprocess(qOut)
